pyproject.toml
[tool.poetry]
name = "postitai"
version = "0.1.0"
description = ""
authors = ["Aditya Karnam <akarnam37@gmail.com>"]
license = "MIT"
readme = "README.md"

[flake8]
per-file-ignores = "util/response.py:F501"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''

[tool.poetry.dependencies]
python = "^3.11"
newscatcherapi = "^0.7.2"
python-dotenv = "^1.0.0"
click = "^8.1.3"
black = "^23.3.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
openai = "^0.27.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
